+++ src/components/Logbook/EventRows.tsx
@@ -1,5 +1,5 @@
 import { FC } from 'react';
-import EventRowsI from '../interfaces/EventRowsI';
+import EventRowsI from '../../interfaces/EventRowsI';
 import LogBookRow from './LogBookRow';
 
 const EventRows: FC<EventRowsI> = (props): JSX.Element => {
