+++ src/components/Logbook/LogBookHistoryRows.tsx
@@ -1,5 +1,5 @@
 import { FC } from 'react';
-import LogBookHistoryRowsProps from '../interfaces/LogBookHistoryRowsProps';
+import LogBookHistoryRowsProps from '../../interfaces/LogBookHistoryRowsProps';
 import LogBookHistoryRowComp from './LogBookHistoryRow';
 
 const LogBookHistoryRows: FC<LogBookHistoryRowsProps> = (
