+++ src/components/UserSettings/DivingSetSettings.tsx
@@ -1,0 +1,7 @@
+export const DivingSetSettings = (): JSX.Element => {
+  return (
+    <div className="pt-5">
+      <h2>Omat pullot</h2>
+    </div>
+  );
+};
