+++ src/components/UserSettings/UserSettings.tsx
@@ -1,0 +1,8 @@
+export const UserSettings = (): JSX.Element => {
+  return (
+    <div>
+      <h2>Omat tiedot</h2>
+      Nimi Puhelinnumero Jotain
+    </div>
+  );
+};
