+++ src/index.css
@@ -1,8 +1,7 @@
 body {
   margin: 0;
-  font-family: 'Roboto', -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto',
-    'Oxygen', 'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
-    sans-serif;
+  font-family: 'Roboto' -apple-system BlinkMacSystemFont 'Segoe UI' 'Roboto' 'Oxygen'
+    'Ubuntu' 'Cantarell' 'Fira Sans' 'Droid Sans' 'Helvetica Neue' sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
 }
@@ -16,8 +15,7 @@ input {
 }
 
 code {
-  font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
-    monospace;
+  font-family: source-code-pro Menlo Monaco Consolas 'Courier New' monospace;
 }
 
 #signInContainer {
